requirements.txt
alembic==1.0.10
asn1crypto==0.24.0
astroid==2.0.4
Babel==2.6.0
bcrypt==3.1.6
blinker==1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.6.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.8
django-appconf==1.0.3
django-bootstrap3==11.0.0
django-compressor==2.2
django-extra-views==0.11.0
django-haystack==2.8.1
django-heroku==0.3.1
django-misaka==0.2.1
django-oscar==1.6.7
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
dodgy==0.1.9
factory-boy==2.12.0
Faker==1.0.7
flake8==3.5.0
flake8-polyfill==1.0.2
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Gravatar==0.5.0
Flask-HTMLmin==1.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
houdini.py==0.1.0
htmlmin==0.1.12
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Mako==1.0.10
MarkupSafe==1.1.1
mccabe==0.6.1
misaka==2.1.0
mock==2.0.0
oauthlib==2.1.0
paramiko==2.4.2
passlib==1.7.1
pbr==5.2.0
pep8-naming==0.7.0
pgadmin4==4.6
phonenumbers==8.10.12
Pillow==5.3.0
prospector==1.1.2
psutil==5.5.1
psycopg2==2.7.4
psycopg2-binary==2.8.2
purl==1.5
pyasn1==0.4.5
pycodestyle==2.3.1
pycparser==2.18
pydocstyle==2.1.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
PyNaCl==1.3.0
python-dateutil==2.8.0
python-decouple==3.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
rcssmin==1.0.6
requests==2.19.1
requests-oauthlib==1.0.0
requirements-detector==0.6
rjsmin==1.0.12
setoptconf==0.2.0
simplejson==3.16.0
six==1.12.0
snowballstemmer==1.2.1
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sorl-thumbnail==12.4.1
South==1.0.2
speaklater==1.3
SQLAlchemy==1.3.3
sqlparse==0.2.4
sshtunnel==0.1.4
text-unidecode==1.2
typed-ast==1.1.0
Unidecode==1.0.23
urllib3==1.23
Werkzeug==0.15.4
whitenoise==4.0
wrapt==1.10.11
WTForms==2.2.1
